+++ frontend/src/components/reader/pager/horizontal-pages.vue
@@ -4,14 +4,12 @@
       v-if="showPageOne"
       ref="pageOne"
       :book="book"
-      :book-settings="bookSettings"
       :page="pageOne"
       class="horizontalPage"
     />
     <BookPage
       v-if="showPageTwo"
       :book="book"
-      :book-settings="bookSettings"
       :page="pageTwo"
       class="horizontalPage"
     />
@@ -30,16 +28,19 @@ export default {
   components: { BookPage, ScaleForScroll },
   props: {
     book: { type: Object, required: true },
-    bookSettings: { type: Object, required: true },
     page: { type: Number, required: true },
-    isReadInReverse: { type: Boolean, required: true },
   },
   emits: ["click"],
   computed: {
+    bookSettings() {
+      return this.getBookSettings(this.book);
+    },
+    isReadInReverse() {
+      return this.bookSettings.isReadInReverse;
+    },
     showSecondPage() {
-      const settings = this.getSettings(this.book);
       return (
-        settings.twoPages &&
+        this.bookSettings.twoPages &&
         this.page < this.book.maxPage &&
         !this.isCoverPage(this.book, this.page)
       );
@@ -58,7 +59,7 @@ export default {
     },
   },
   methods: {
-    ...mapActions(useReaderStore, ["getSettings", "isCoverPage"]),
+    ...mapActions(useReaderStore, ["getBookSettings", "isCoverPage"]),
   },
 };
 </script>