+++ frontend/src/components/reader/pager/pager-horizontal.vue
@@ -20,12 +20,7 @@
       :model-value="page"
       :transition="true"
     >
-      <HorizontalPages
-        :book="book"
-        :book-settings="bookSettings"
-        :page="page"
-        :is-read-in-reverse="isReadInReverse"
-      />
+      <HorizontalPages :book="book" :page="page" />
     </v-window-item>
   </v-window>
 </template>
@@ -35,7 +30,7 @@ import { mapActions, mapState } from "pinia";
 
 import HorizontalPages from "@/components/reader/pager/horizontal-pages.vue";
 import PageChangeLink from "@/components/reader/pager/page-change-link.vue";
-import { REVERSE_READING_DIRECTIONS, useReaderStore } from "@/stores/reader";
+import { useReaderStore } from "@/stores/reader";
 import { range } from "@/util";
 
 const WINDOW_BACK_BOUND = 48;
@@ -49,7 +44,6 @@ export default {
   },
   props: {
     book: { type: Object, required: true },
-    bookSettings: { type: Object, required: true },
   },
   data() {
     return {
@@ -64,11 +58,14 @@ export default {
       storePage: (state) => state.page,
       storePk: (state) => state.books.current.pk,
     }),
+    bookSettings() {
+      return this.getBookSettings(this.book);
+    },
     twoPages() {
       return this.bookSettings.twoPages;
     },
     isReadInReverse() {
-      return REVERSE_READING_DIRECTIONS.has(this.bookSettings.readingDirection);
+      return this.bookSettings.isReadInReverse;
     },
     windowIndex() {
       const val = this.activePage - this.pages[0];
@@ -114,7 +111,11 @@ export default {
     this.setPages();
   },
   methods: {
-    ...mapActions(useReaderStore, ["setBookChangeFlag", "setActivePage"]),
+    ...mapActions(useReaderStore, [
+      "getBookSettings",
+      "setBookChangeFlag",
+      "setActivePage",
+    ]),
     setPages() {
       const backPages = Math.max(this.activePage - WINDOW_BACK_BOUND, 0);
       const forePages = Math.min(